feat: update enums
