+++ src/database/index.js
@@ -4,7 +4,7 @@ const router = jsonServer.router('db.json')
 const middlewares = jsonServer.defaults()
 const localIpAddress = require('local-ip-address')
 const Logger = require('@ptkdev/logger')
-const port = process.env.PORT || 3001 // you can use any port number here; i chose to use 3001
+const port = process.env.PORT || 3001 // use any port number here
 const logger = new Logger()
 
 server.use(middlewares)
