+++ src/scripts/constants/enums.ts
@@ -1,0 +1,26 @@
+export enum apiMethods {
+    Get = 'GET',
+    Post = 'POST',
+    Put = 'PUT',
+    Patch = 'PATCH',
+    Delete = 'DELETE',
+}
+
+export enum inValidGenreFields {
+    Empty,
+    Repeated,
+}
+
+// used to detemine the which key typed on the keyboard
+export enum EVENT_CODE {
+    ENTER = 'Enter',
+    ESC = 'Escape',
+    TAB = 'Tab',
+}
+
+// used to define type of the modal to render
+export const MODAL_TYPE = {
+    SONG_DETAIL: 'Song detail',
+    ADD_SONG: 'Add new song',
+    EDIT_SONG: 'Edit song',
+}
